+++ packages/lib-user/src/components/MyGroups/components/GroupCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './GroupCard'
